+++ ldap/servers/plugins/replication/repl_connext.c
@@ -111,7 +111,7 @@ void consumer_connection_extension_destructor (void *ext, void *object, void *pa
 					slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 									"Aborting total update in progress for replicated "
 									"area %s connid=%" NSPRIu64 "\n", slapi_sdn_get_dn(repl_root_sdn),
-									connid);
+									(long long unsigned int)connid);
 					slapi_stop_bulk_import(pb);
 				}
 				else
@@ -183,20 +183,20 @@ consumer_connection_extension_acquire_exclusive_access(void* conn, PRUint64 conn
             ret = connext;
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Acquired consumer connection extension\n",
-                    connid, opid);
+                    (long long unsigned int)connid, opid);
         }
         else if (opid == connext->in_use_opid)
         {
             ret = connext;
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Reacquired consumer connection extension\n",
-                    connid, opid);
+                    (long long unsigned int)connid, opid);
         }
         else
         {
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Could not acquire consumer connection extension; it is in use by op=%d\n",
-                    connid, opid, connext->in_use_opid);
+                    (long long unsigned int)connid, opid, connext->in_use_opid);
         }
 
         /* step 5, drop the lock */
@@ -206,7 +206,7 @@ consumer_connection_extension_acquire_exclusive_access(void* conn, PRUint64 conn
     {
         slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                 "conn=%" NSPRIu64 " op=%d Could not acquire consumer extension, it is NULL!\n",
-                connid, opid);
+                (long long unsigned int)connid, opid);
     }
     
     return ret;
@@ -244,7 +244,7 @@ consumer_connection_extension_relinquish_exclusive_access(void* conn, PRUint64 c
         {
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Consumer connection extension is not in use\n",
-                    connid, opid);
+                    (long long unsigned int)connid, opid);
             ret = 2;
         }
         else if (opid == connext->in_use_opid)
@@ -252,7 +252,7 @@ consumer_connection_extension_relinquish_exclusive_access(void* conn, PRUint64 c
             /* step 4, relinquish it (normal) */
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Relinquishing consumer connection extension\n",
-                    connid, opid);            
+                    (long long unsigned int)connid, opid);
             connext->in_use_opid = -1;
             ret = 1;
         }
@@ -261,7 +261,7 @@ consumer_connection_extension_relinquish_exclusive_access(void* conn, PRUint64 c
             /* step 4, relinquish it (forced) */
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Forced to relinquish consumer connection extension held by op=%d\n",
-                    connid, opid, connext->in_use_opid);
+                    (long long unsigned int)connid, opid, connext->in_use_opid);
             connext->in_use_opid = -1;
             ret = 1;
         }
@@ -269,7 +269,7 @@ consumer_connection_extension_relinquish_exclusive_access(void* conn, PRUint64 c
         {
             slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                     "conn=%" NSPRIu64 " op=%d Not relinquishing consumer connection extension, it is held by op=%d!\n",
-                    connid, opid, connext->in_use_opid);         
+                    (long long unsigned int)connid, opid, connext->in_use_opid);
         }
         
         /* step 5, drop the lock */
@@ -279,7 +279,7 @@ consumer_connection_extension_relinquish_exclusive_access(void* conn, PRUint64 c
     {
         slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
                 "conn=%" NSPRIu64 " op=%d Could not relinquish consumer extension, it is NULL!\n",
-                connid, opid);
+                (long long unsigned int)connid, opid);
     }
     
     return ret;
