+++ ldap/servers/slapd/back-ldbm/ldbm_delete.c
@@ -122,7 +122,8 @@ ldbm_back_delete( Slapi_PBlock *pb )
 
 	if (pb->pb_conn)
 	{
-		slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_delete", "enter conn=%" NSPRIu64 " op=%d\n", pb->pb_conn->c_connid, operation->o_opid);
+		slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_delete", "enter conn=%" NSPRIu64 " op=%d\n",
+				(long long unsigned int)pb->pb_conn->c_connid, operation->o_opid);
 	}
 
 	if ((NULL == addr) || (NULL == sdnp))
@@ -1284,7 +1285,8 @@ diskfull_return:
 	slapi_ch_free_string(&e_uniqueid);
 	if (pb->pb_conn)
 	{
-		slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_delete", "leave conn=%" NSPRIu64 " op=%d\n", pb->pb_conn->c_connid, operation->o_opid);
+		slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_delete", "leave conn=%" NSPRIu64 " op=%d\n",
+				(long long unsigned int)pb->pb_conn->c_connid, operation->o_opid);
 	}
 	return rc;
 }
